Used fixed-with integral type to match the one in the LLVM pass

+++ tracer/tracerlib.c
@@ -7,10 +7,10 @@
 #define LOAD 0
 #define STORE 1
 
-void loopEvent(int type);
-void memoryEvent(int type, void *addr, uint64_t pc);
+void loopEvent(uint8_t type);
+void memoryEvent(uint8_t type, void *addr, uint64_t pc);
 
-void loopEvent(int type) {
+void loopEvent(uint8_t type) {
   switch (type) {
     case LOOP_ENTRY: printf("LoopEntry\n"); break;
     case LOOP_ITER: printf("LoopIter\n"); break;
@@ -19,7 +19,7 @@ void loopEvent(int type) {
   }
 }
 
-void memoryEvent(int type, void *addr, uint64_t pc) {
+void memoryEvent(uint8_t type, void *addr, uint64_t pc) {
   switch (type) {
     case LOAD: printf("Load %llu %p\n", pc, addr); break;
     case STORE: printf("Store %llu %p\n", pc, addr); break;
